Proposal: Add support for WPCOM Geo Uniques
VIP Go Geo Uniques is our VIP Go port of the WPCOM Geo Uniques plugin. The function names are changed, but the arguments are the same. This will make it easier to copy VIP themes as-is from WordPress.com.

Keeping differences as minimal as possible while we're keeping the themes in sync across platforms will ease the migration process and then we can update the function names after everything is fully migrated.